tooling: re-enable bench_util (#10674)
