+++ bench_util/Cargo.toml
@@ -1,7 +1,7 @@
 # Copyright 2018-2021 the Deno authors. All rights reserved. MIT license.
 [package]
-name = "bench_util"
-version = "0.0.0"
+name = "deno_bench_util"
+version = "0.1.0"
 authors = ["the Deno authors"]
 edition = "2018"
 description = "Bench and profiling utilities for deno crates"
@@ -16,3 +16,7 @@ publish = false
 bencher = "0.1"
 deno_core = { version = "0.88.0", path = "../core" }
 tokio = { version = "1.6.0", features = ["full"] }
+
+[[bench]]
+name = "op_baseline"
+harness = false
