+++ extensions/url/Cargo.toml
@@ -20,7 +20,7 @@ percent-encoding = "2.1.0"
 serde = { version = "1.0.125", features = ["derive"] }
 
 [dev-dependencies]
-bencher = "0.1"
+deno_bench_util = { version = "0.1.0", path = "../../bench_util" }
 
 [[bench]]
 name = "url_ops"
